# Minor optimization for accumulator types in the parser