I had a [TODO](https://github.com/bufbuild/protocompile/blob/99bab3cee7f30bf070855a494d5a1e1ec6a497f5/parser/ast.go#L22) in the code that I thought might slightly improve the perf/memory usage in here.

Before addresssing the TODO, I made a small micro-benchmark to test a few different approaches:
```
BenchmarkSlicePointerForEachElement-10       	  994368	      1197 ns/op	    2303 B/op	      35 allocs/op
BenchmarkSliceReallocAtEnd_Exact-10          	 1543438	       790.3 ns/op	    2226 B/op	      12 allocs/op
BenchmarkSliceReallocAtEnd_Within10Pct-10    	 1457452	       793.3 ns/op	    2190 B/op	      12 allocs/op
BenchmarkSliceReallocAtEnd_Within20Pct-10    	 1635313	       771.7 ns/op	    2137 B/op	      12 allocs/op
BenchmarkSliceReallocAtEnd_Within25Pct-10    	 1714380	       757.8 ns/op	    2120 B/op	      12 allocs/op
BenchmarkSliceReallocAtEnd_Within30Pct-10    	 1625214	       758.3 ns/op	    2105 B/op	      12 allocs/op
BenchmarkSlice-10                            	 2048469	       590.3 ns/op	    1658 B/op	      11 allocs/op
BenchmarkLinkedList-10                       	 1360090	       878.0 ns/op	    1383 B/op	      26 allocs/op
PASS
Untruncated slices used 44.3% more capacity than needed
```
Many things in the grammar are already using the slice approach (which, per above, is by far the fastest). But many things were using linked lists instead.

Here's what the benchmark scenarios mean:
* `LinkedList`: Uses a linked list and accumulates the list from right-to-left (tail first), so that each step is just pushing a new head of the list (similar to how linked list is used to push items on a treiber stack). The last step here is to convert the list to a slice. The capacity of the slice is allocated precisely, since we know how many items are in the list. (The cost of that alloc and copying are in the above benchmarks.)
* `Slice`: The most intuitive approach, which accumulates from left-to-right, appending each new element. This turned out to be by far the fastest, but at the cost of using 40-50% more capacity than needed (so the final model uses more memory than with the linked list approach).
* `SliceReallocAtEnd`: This is exactly like `Slice`, except that at the end, it creates a new slice that has exactly the capacity needed and copies into it. This is slower since it must do an extra slice alloc-and-copy at the end, but it attempts to improve on the total memory used by the final model/AST. There are several scenarios in the benchmarks, to see if different thresholds had an impact on performance. The thresholds mean that the logic only does the copy when the final slice's capacity is otherwise too large (e.g. >25% of the needed capacity). Surprisingly, even being tolerant up to 30% overage didn't really improve the performance much.
* `SlicePointerForEachElement`: This is the absolute worst benchmark. It incurs the cost of a linked list, in that it must perform an allocation for every element. But it also must do the incremental slice re-allocation for growing slices.

So I changed **nearly all** of the things using linked lists to instead use slices. There is one remaining linked list:the way field+value pairs are accumulated for message literals. We could try to complicate the grammar to avoid needing a new item for each element. But that reduces clarity of the grammar. So we keep it as a linked list, which is still much faster than the alternative (`SlicePointerForEachElement`).

In the end, this didn't meaningfully make any impact on the benchmarks. Which is a bit unfortunate but not really unexpected: the parse step is, overall, much cheaper than the link step. However, I think it's still worth considering merging anyway since it makes accumulation (other than the one case for message literals) uniform and consistent (and simpler).

The only likely way to make the parse step materially faster is to parse _directly_ to descriptor protos instead of to a real AST. The old protoparse package did that once upon a time. The main downside is that error messages when interpreting options are worse, due to much less precise information about elements inside message literals. Also, we _need_ the AST for certain things, like formatting, since it is non-lossy. So if we do that, we actually would need to do a couple of things (1) support _both_ (so can still get an AST for things that need it, but skip creating the AST when it's otherwise not needed/used) and (2) maintain parsing into an AST for message literals, even when directly building descriptors, so that we have both available for good error messages and position information when interpreting options.